fix: moved all-components inside component folder
