+++ src/pages/components/index.mdx
@@ -1,3 +1,3 @@
-import Components from "@/all-components/Components";
+import Components from "@/components/all-components/Components";
 
 <Components />
